refactor: splitted ExerciseInfoModal in multiple components
