+++ src/components/Card/OerCard/ResourceCardsList.tsx
@@ -115,12 +115,12 @@ export default function ResourceCardsList({
                               ? isResourcePage && collectionsColor[0]
                                 ? collectionsColor[0]
                                 : // : collectionsColor[   // to handle when we use the API with pagination
-                                // (currentPage) > 1
-                                //   ? index + itemsPerPage * (currentPage - 1)
-                                //   : index]
-                                collectionsColor[index]
+                                  // (currentPage) > 1
+                                  //   ? index + itemsPerPage * (currentPage - 1)
+                                  //   : index]
+                                  collectionsColor[index]
                               : //: collectionsColor[index] //this is the logic to color the iconBookmark of each card with the right color. Without this logic, the color of the iconBookmark is always only the first #itemsPerPage colors of the collectionsColor array
-                              ''
+                                ''
                           }
                           // oer={fill_the_gaps_data}
                           oer={oer}
@@ -147,8 +147,8 @@ export default function ResourceCardsList({
                             //alert("Non rispettato il primo if \n collectionIndex: " + collectionIndex)
                             //}
                           }}
-                        //position="absolute"
-                        //right={'0px'}
+                          //position="absolute"
+                          //right={'0px'}
                         >
                           <DeleteIcon />
                         </Button>
@@ -233,8 +233,8 @@ export default function ResourceCardsList({
                             //alert("Non rispettato il primo if \n collectionIndex: " + collectionIndex)
                             //}
                           }}
-                        //position="absolute"
-                        //right={'0px'}
+                          //position="absolute"
+                          //right={'0px'}
                         >
                           <DeleteIcon />
                         </Button>
