+++ src/types/encoreElements/ExerciseInfoModalProps.ts
@@ -1,11 +1,11 @@
 import { ColorCollectionProps } from './collectionModal';
 
-export type ExerciseInfoModalProps = {
-  isOpen: boolean;
-  onClose: () => void;
-  title: string;
-  authors: (string | null)[];
-  coverage: string[];
+export interface ExerciseInfoModalProps {
+  isOpen?: boolean;
+  onClose?: () => void;
+  title?: string;
+  authors?: (string | null)[];
+  coverage?: string[];
   fill_template?: string | null;
   fill_template_with_gaps?: string | null;
   n_o_w?: number | null;
@@ -20,10 +20,10 @@ export type ExerciseInfoModalProps = {
   type_of_exercise?: string | null; // This is used only for the multiple choice exercises
   type_of_question?: string | null; // This is used only for the open question exercises
   category?: string | null;
-  showTagDigital: boolean;
-  showTagEntrepreneurial: boolean;
-  showTagGreen: boolean;
-  isGeneratedByAI: boolean;
+  showTagDigital?: boolean;
+  showTagEntrepreneurial?: boolean;
+  showTagGreen?: boolean;
+  isGeneratedByAI?: boolean;
   collectionsColor?: (ColorCollectionProps | undefined)[];
   assessment_oer_type?: string | null;
-};
+}
